+++ VpnHood.Common/Exceptions/AnotherInstanceIsRunning.cs
@@ -1,0 +1,10 @@
+﻿using System;
+
+namespace VpnHood.Common.Exceptions;
+
+public class AnotherInstanceIsRunning : Exception
+{
+    public AnotherInstanceIsRunning() : base("Another instance is running.")
+    {
+    }
+}